Material Dialog y snackbar y mejoras

+++ src/app/heroes/views/add-heroe/add-heroe.component.ts
@@ -3,6 +3,9 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { Heroe, Publisher } from '../../model/heroe.model';
 import { HeroesService } from '../../services/heroes.service';
 import { switchMap } from 'rxjs/operators';
+import { MatSnackBar } from '@angular/material/snack-bar';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmarComponent } from '../../components/confirmar/confirmar.component';
 
 @Component({
   selector: 'app-add-heroe',
@@ -35,7 +38,9 @@ export class AddHeroeComponent implements OnInit {
   constructor(
     private heroesService: HeroesService,
     private activatedRoute: ActivatedRoute,
-    private router: Router
+    private router: Router,
+    private snackBar: MatSnackBar,
+    public dialog: MatDialog
   ) {}
 
   ngOnInit(): void {
@@ -59,20 +64,40 @@ export class AddHeroeComponent implements OnInit {
     if (this.heroe.id) {
       this.heroesService
         .updateHeroe(this.heroe)
-        .subscribe((heroe) => console.log('Actualizando', heroe));
+        .subscribe((heroe) => this.mostrarSnackbar('Registro actualizado'));
     }
     // si no hay id, creo uno nuevo
     else {
       this.heroesService.addHeroe(this.heroe).subscribe((heroe) => {
         this.router.navigate(['/heroes', heroe.id]);
+        this.mostrarSnackbar('Registro creado');
       });
     }
   }
 
   // Borrar héroe
-  borrar(): void {
-    this.heroesService.deleteHeroe(this.heroe.id).subscribe((resp) => {
-      this.router.navigate(['/heroes']);
+  borrarHeroe(): void {
+    //MatDialog
+    const dialog = this.dialog.open(ConfirmarComponent, {
+      width: '250px',
+      // como los obj son pasados por ref no queremos que se modifique ya que el dialog solo debe ser para lectura
+      data: { ...this.heroe },
+    });
+
+    dialog.afterClosed().subscribe((result) => {
+      if (result) {
+        this.heroesService.deleteHeroe(this.heroe.id).subscribe((resp) => {
+          this.router.navigate(['/heroes']);
+          this.mostrarSnackbar('Registro borrado');
+        });
+      }
+    });
+  }
+
+  // SnackBar recibe como arg el mensaje que quiero mostrar
+  mostrarSnackbar(mensaje: string): void {
+    this.snackBar.open(mensaje, 'Cerrar', {
+      duration: 2500,
     });
   }
 }
